requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
transformers>=4.52.0
torch>=2.0.0
accelerate>=0.21.0
safetensors>=0.3.0
pydantic==2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-community>=0.3.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0

